Move Google Map files to a folder under AboutUs.
